requirements.txt
pytest==7.2.2
pytest-xdist==3.2.1
selenium==4.8.3
allure-pytest==2.13.1
faker==18.4.0
